feat: aggregate new models.rs module, closes #20
